Gates

+++ app/Http/Controllers/NewsController.php
@@ -5,7 +5,9 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\Gate;
 use App\News;
+use DB;
 use App\Http\Requests\StoreNews;
 use App\Http\Requests\UpdateNews;
 use App\Services\PivotService;
@@ -47,27 +49,47 @@ class NewsController extends Controller
     {
         $this->news->setTable(Auth::User()->prefix.'_news');
         $news = $this->news->get()->where('id',$request->id);
-        return response()->json([
-            'message' => 'Success, found data!',
-            'data' => $news
-        ], 201);  
+
+        if (Gate::allows('edit-news', $news)) {
+            return response()->json([
+                'message' => 'Success, found data!',
+                'data' => $news
+            ], 201);  
+        }else{
+            return response()->json([
+                'message' => 'It is not your news!'], 403);
+        }
     }
 
     public function update(UpdateNews $request)
     {   
+        $news = DB::table(Auth::User()->prefix.'_news')->where('id', $request->id)->first();
+        //$this->authorize('update',Auth::user(), $request);
+        if (Gate::allows('update-news', $news)) {
 
-        $this->news->setTable(Auth::User()->prefix.'_news');
-        $this->news->where('id',$request->id)->update($request->validated());
+            $this->news->setTable(Auth::User()->prefix.'_news');
+            $this->news->where('id',$request->id)->update($request->validated());
+            return response()->json([
+                'message' => 'Success, data updated!'], 201);
 
-        return response()->json([
-            'message' => 'Success, data updated!'], 201);
+          }else{
+            return response()->json([
+                'message' => 'It is not your news!'], 403);
+          }
     }
 
     public function delete(Request $request)
     {
-        $this->news->setTable(Auth::User()->prefix.'_news');
-        $this->news->where('id',$request->id)->delete();
-        return response()->json([
-            'message' => 'Success, data deleted'], 201);
+        $news = DB::table(Auth::User()->prefix.'_news')->where('id', $request->id)->first();
+        
+        if (Gate::allows('delete-news', $news)) {
+            $this->news->setTable(Auth::User()->prefix.'_news');
+            $this->news->where('id',$request->id)->delete();
+            return response()->json([
+                'message' => 'Success, data deleted'], 201);
+        }else{
+            return response()->json([
+                'message' => 'It is not your news!'], 403);
+        }
     }
 }
